+++ packages/expect/src/index.ts
@@ -412,6 +412,9 @@ setMatchers(spyMatchers, true, expect as Expect);
 setMatchers(toThrowMatchers, true, expect as Expect);
 
 expect.addSnapshotSerializer = () => void 0;
+expect.deadline = () => {
+  throw new Error('deadline must be implemented by the runtime');
+};
 expect.assertions = assertions;
 expect.hasAssertions = hasAssertions;
 expect.getState = getState;