+++ packages/jest-circus/src/run.ts
@@ -150,6 +150,7 @@ const _callCircusHook = async ({
 }): Promise<void> => {
   await dispatch({hook, name: 'hook_start'});
   const timeout = hook.timeout || getState().testTimeout;
+  updateDeadline(timeout);
 
   try {
     await callAsyncCircusFn(hook, testContext, {
@@ -168,6 +169,7 @@ const _callCircusTest = async (
 ): Promise<void> => {
   await dispatch({name: 'test_fn_start', test});
   const timeout = test.timeout || getState().testTimeout;
+  updateDeadline(timeout);
   invariant(test.fn, `Tests with no 'fn' should have 'mode' set to 'skipped'`);
 
   if (test.errors.length) {
@@ -184,5 +186,9 @@ const _callCircusTest = async (
     await dispatch({error, name: 'test_fn_failure', test});
   }
 };
+
+function updateDeadline(timeout: number): void {
+  getState().currentlyRunningChildDeadline = Date.now() + timeout - 20;
+}
 
 export default run;